Various fixes and features implemented during doc development
* Better messaging for expiration.
* Fixed already stopped message.
* retention-archive and retention-archive-type now use retention-full and 'full' when not specified.
* Fixed issue where backup-user was required (should default to backrest).
* ExecuteTest now supports retries.
* Fixed issue where log test was not comparing test logs.
* Fixed issue where test logs would not match for ssh connection errors

+++ lib/BackRest/Expire.pm
@@ -125,27 +125,35 @@ sub process
         # Make sure iFullRetention is valid
         if (!looks_like_number($iFullRetention) || $iFullRetention < 1)
         {
-            confess &log(ERROR, 'full_retention must be a number >= 1');
+            confess &log(ERROR, 'retention-full must be a number >= 1');
         }
 
-        my $iIndex = $iFullRetention;
-        @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(true), 'reverse');
+        @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(true));
 
-        while (defined($stryPath[$iIndex]))
+        if (@stryPath > $iFullRetention)
         {
             # Delete all backups that depend on the full backup.  Done in reverse order so that remaining backups will still
             # be consistent if the process dies
-            foreach $strPath ($oFile->list(PATH_BACKUP_CLUSTER, undef, '^' . $stryPath[$iIndex] . '.*', 'reverse'))
+            for (my $iFullIdx = 0; $iFullIdx < @stryPath - $iFullRetention; $iFullIdx++)
             {
-                system("rm -rf ${strBackupClusterPath}/${strPath}") == 0
-                    or confess &log(ERROR, "unable to delete backup ${strPath}");
+                my @stryRemoveList;
 
-                $oBackupInfo->delete($strPath);
+                foreach $strPath ($oFile->list(PATH_BACKUP_CLUSTER, undef, '^' . $stryPath[$iFullIdx] . '.*'))
+                {
+                    system("rm -rf ${strBackupClusterPath}/${strPath}") == 0
+                        or confess &log(ERROR, "unable to delete backup ${strPath}");
+
+                    $oBackupInfo->delete($strPath);
+
+                    if ($strPath ne $stryPath[$iFullIdx])
+                    {
+                        push(@stryRemoveList, $strPath);
+                    }
+                }
+
+                &log(INFO, 'remove expired full backup ' . (@stryRemoveList > 0 ? 'set: ' : '')  . $stryPath[$iFullIdx] .
+                           (@stryRemoveList > 0 ? ', ' . join(', ', @stryRemoveList) : ''));
             }
-
-            &log(INFO, 'remove expired full backup: ' . $stryPath[$iIndex]);
-
-            $iIndex++;
         }
     }
 
@@ -155,30 +163,58 @@ sub process
         # Make sure iDifferentialRetention is valid
         if (!looks_like_number($iDifferentialRetention) || $iDifferentialRetention < 1)
         {
-            confess &log(ERROR, 'differential_retention must be a number >= 1');
+            confess &log(ERROR, 'retention-diff must be a number >= 1');
         }
 
-        @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(false, true), 'reverse');
+        @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(false, true));
 
-        if (defined($stryPath[$iDifferentialRetention - 1]))
+        if (@stryPath > $iDifferentialRetention)
         {
-            logDebugMisc($strOperation,  'differential expiration based on ' . $stryPath[$iDifferentialRetention - 1]);
+            # $strDiffRetain = $stryPath[@stryPath - $iDifferentialRetention];
 
-            # Get a list of all differential and incremental backups
-            foreach $strPath ($oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(false, true, true), 'reverse'))
+            # logDebugMisc($strOperation,  'differential expiration based on ' . $strDiffRetain);
+
+            for (my $iDiffIdx = 0; $iDiffIdx < @stryPath - $iDifferentialRetention; $iDiffIdx++)
             {
-                logDebugMisc($strOperation, "checking ${strPath} for differential expiration");
+                # Get a list of all differential and incremental backups
+                my @stryRemoveList;
 
-                # Remove all differential and incremental backups before the oldest valid differential
-                if ($strPath lt $stryPath[$iDifferentialRetention - 1])
+                foreach $strPath ($oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(false, true, true)))
                 {
-                    system("rm -rf ${strBackupClusterPath}/${strPath}") == 0
-                        or confess &log(ERROR, "unable to delete backup ${strPath}");
-                    $oBackupInfo->delete($strPath);
+                    logDebugMisc($strOperation, "checking ${strPath} for differential expiration");
 
-                    &log(INFO, "remove expired diff/incr backup ${strPath}");
+                    # Remove all differential and incremental backups before the oldest valid differential
+                    if ($strPath lt $stryPath[$iDiffIdx + 1])
+                    {
+                        system("rm -rf ${strBackupClusterPath}/${strPath}") == 0
+                            or confess &log(ERROR, "unable to delete backup ${strPath}");
+                        $oBackupInfo->delete($strPath);
+
+                        if ($strPath ne $stryPath[$iDiffIdx])
+                        {
+                            push(@stryRemoveList, $strPath);
+                        }
+                        # &log(INFO, "remove expired diff/incr backup ${strPath}");
+                    }
                 }
+
+                &log(INFO, 'remove expired diff backup ' . (@stryRemoveList > 0 ? 'set: ' : '')  . $stryPath[$iDiffIdx] .
+                           (@stryRemoveList > 0 ? ', ' . join(', ', @stryRemoveList) : ''));
             }
+        }
+    }
+
+    # Default archive retention if not explicily set
+    if (defined($iFullRetention))
+    {
+        if (!defined($iArchiveRetention))
+        {
+            $iArchiveRetention = $iFullRetention;
+        }
+
+        if (!defined($strArchiveRetentionType))
+        {
+            $strArchiveRetentionType = BACKUP_TYPE_FULL;
         }
     }
 
@@ -192,20 +228,10 @@ sub process
         # Determine which backup type to use for archive retention (full, differential, incremental)
         if ($strArchiveRetentionType eq BACKUP_TYPE_FULL)
         {
-            if (!defined($iArchiveRetention))
-            {
-                $iArchiveRetention = $iFullRetention;
-            }
-
             @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(true), 'reverse');
         }
         elsif ($strArchiveRetentionType eq BACKUP_TYPE_DIFF)
         {
-            if (!defined($iArchiveRetention))
-            {
-                $iArchiveRetention = $iDifferentialRetention;
-            }
-
             @stryPath = $oFile->list(PATH_BACKUP_CLUSTER, undef, backupRegExpGet(true, true), 'reverse');
         }
         elsif ($strArchiveRetentionType eq BACKUP_TYPE_INCR)
@@ -228,7 +254,7 @@ sub process
             confess &log(ERROR, 'archive_retention must be a number >= 1');
         }
 
-        # if no backups were found then preserve current archive logs - too scary to delete them!
+        # if no backups were found then preserve current archive logs - too soon to expire them
         my $iBackupTotal = scalar @stryPath;
 
         if ($iBackupTotal > 0)
@@ -240,8 +266,7 @@ sub process
             {
                 if ($strArchiveRetentionType eq BACKUP_TYPE_FULL && scalar @stryPath > 0)
                 {
-                    &log(INFO, 'fewer than required backups for retention, ' .
-                               'but since archive_retention_type = full using oldest full backup');
+                    &log(INFO, "full backup total < ${iArchiveRetention} - using oldest full backup for archive retention");
                     $strArchiveRetentionBackup = $stryPath[scalar @stryPath - 1];
                 }
             }
@@ -253,8 +278,6 @@ sub process
 
                 # Get the archive logs that need to be kept.  To be cautious we will keep all the archive logs starting from this
                 # backup even though they are also in the pg_xlog directory (since they have been copied more than once).
-                &log(INFO, 'archive retention based on backup ' . $strArchiveRetentionBackup);
-
                 my $oManifest = new BackRest::Manifest($oFile->pathGet(PATH_BACKUP_CLUSTER) .
                                                        "/${strArchiveRetentionBackup}/backup.manifest");
                 my $strArchiveLast = $oManifest->get(MANIFEST_SECTION_BACKUP, MANIFEST_KEY_ARCHIVE_START);
@@ -264,7 +287,7 @@ sub process
                     confess &log(ERROR, "invalid archive location retrieved ${strArchiveRetentionBackup}");
                 }
 
-                &log(INFO, 'archive retention starts at ' . $strArchiveLast);
+                &log(INFO, "archive retention from backup ${strArchiveRetentionBackup}, start = ${strArchiveLast}");
 
                 # Get archive info
                 my $oArchive = new BackRest::Archive();
@@ -326,7 +349,8 @@ sub process
                 }
                 else
                 {
-                    &log(INFO, "expired WAL segments: start = ${strArchiveExpireStart}, stop = ${strArchiveExpireStop}");
+                    &log(INFO, 'expire WAL segments: start = ' . substr($strArchiveExpireStart, 0, 24) .
+                               ', stop = ' . substr($strArchiveExpireStop, 0, 24));
                 }
             }
         }
